requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Cython==0.29.21
idna==2.10
numpy==1.19.0
pandas==1.0.5
pyTelegramBotAPI==3.7.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tabulate==0.8.7
urllib3==1.25.9
vedis==0.7.1
